+++ engine/flink/management/src/main/scala/pl/touk/nussknacker/engine/management/FlinkDeploymentManager.scala
@@ -24,17 +24,18 @@ abstract class FlinkDeploymentManager(modelData: BaseModelData, shouldVerifyBefo
 
   private lazy val verification = new FlinkProcessVerifier(modelData.asInvokableModelData)
 
+  override def validate(processVersion: ProcessVersion, deploymentData: DeploymentData, canonicalProcess: CanonicalProcess): Future[Unit] = {
+    checkOldJobStatus(processVersion, canonicalProcess).map(_ => ())
+  }
+
   override def deploy(processVersion: ProcessVersion, deploymentData: DeploymentData, canonicalProcess: CanonicalProcess, savepointPath: Option[String]): Future[Option[ExternalDeploymentId]] = {
     val processName = processVersion.processName
 
     val stoppingResult = for {
-      oldJob <- OptionT(findJobStatus(processName))
+      oldJob <- OptionT(checkOldJobStatus(processVersion, canonicalProcess))
       deploymentId <- OptionT.fromOption[Future](oldJob.deploymentId)
-      _ <- OptionT[Future, Unit](if (!oldJob.allowedActions.contains(ProcessActionType.Deploy))
-        Future.failed(new IllegalStateException(s"Job ${processName.value} cannot be deployed, status: ${oldJob.status.name}")) else Future.successful(Some(())))
       //when it's failed we don't need savepoint...
       if oldJob.isDeployed
-      _ <- OptionT(checkRequiredSlotsExceedAvailableSlots(canonicalProcess, Some(deploymentId)).map(Option(_)))
       maybeSavePoint <- OptionT.liftF(stopSavingSavepoint(processVersion, deploymentId, canonicalProcess))
     } yield {
       logger.info(s"Deploying $processName. Saving savepoint finished")
@@ -52,6 +53,16 @@ abstract class FlinkDeploymentManager(modelData: BaseModelData, shouldVerifyBefo
         savepointPath.orElse(maybeSavepoint)
       )
     } yield runResult
+  }
+
+  private def checkOldJobStatus(processVersion: ProcessVersion, canonicalProcess: CanonicalProcess): Future[Option[ProcessState]] = {
+    val processName = processVersion.processName
+    for {
+      oldJob <- findJobStatus(processName)
+      _ <- if (oldJob.exists(!_.allowedActions.contains(ProcessActionType.Deploy)))
+        Future.failed(new IllegalStateException(s"Job ${processName.value} cannot be deployed, status: ${oldJob.map(_.status.name).getOrElse("")}")) else Future.successful(Some(()))
+      _ <- checkRequiredSlotsExceedAvailableSlots(canonicalProcess, oldJob.flatMap(_.deploymentId))
+    } yield oldJob
   }
 
   protected def checkRequiredSlotsExceedAvailableSlots(canonicalProcess: CanonicalProcess, currentlyDeployedJobId: Option[ExternalDeploymentId]): Future[Unit]